Divide home page components into separate files
